User set cwd, IDD error, BI shading
User sets cwd in plugin, path is passed to this script, and idfs are outputted to this path

IDD may be set in the main script, so this avoids an error if this has already been done

Converts all other BIs to shading. Does not produce an idf if the entire BI is composed of shading blocks.