Shifted results code to fn for layer adding ability
Moved all the code which deals with the results into the add_new_layer function. Added a mode to this fn so that a layer can be added before plugin is run which contains fields to be populated by the user. This is not ideal and would be better to separate result code from layer adding code.