Quick DB bug workaround for testing
Implemented quick workaround for an issue on Mac where certain database attributes would be interpreted as bools and therefore replace 'No' with '0'. These values are replaced after loading the DB files, but the ideal solution would be to define the type at the point of loading. This did not occur on Windows.