Converts DB to idf but needs tweaks
This formed the basis for the functions in the QGIS plugin which converts from the database files into the basic settings idf. This script needs some updates to be fully functional. The version in the plugin is fully functional but needed to be coded slightly different from the standalone script to be integrated. Since it is the intention for Simstock to use a database in the future, we will need this standalone script.