Fixed multipolygon fn bug and removed raise in BI fn
There was a bug in the multipolygon fn which was fixed for the plugin and copying those changes here

The BI function checks after completing whether the adjacent polygons it identified matches with Simstock's internally calculated list. This is producing errors due to two different bugs - investigating. Commented out for now.