Included shading overhang as possible measure
Shading overhang has been added to the Simstock idf geometry script and as an expected attribute for input data. The user can provide any int or float value to specify the depth of the overhang, and this will be added to all windows belonging to the given polygon. If zero or null is provided, no overhang is added.