All fields can be auto-added now
Moved headings to class variable. The Add Fields fn can now add all required fields for Simstock to work. The first time this is done, it will add the general Simstock attributes. After this, if the user populates the nofloors field, the floor-specific fields will be added (needs to be decided).