Updated Mac initial setup lines
The initial setup would fail if run for a second time even if it was successful the first time. This is because the sh script returned status 1. Code has been changed to not raise this as an error but just warn.

Also added a clause to raise an error if ReadVarsESO fails, which was not present before.