Use csvt file to force QGIS field type
Fixed issue of QGIS automatically interpreting field types when importing from csv. Use csvt file to specify the type of each column.

It seems to not care if the number of columns are different to the csv; it will just use as many that are specified and ignore any extra. This is probably only necessary for the schedules csv?