Fixed missing EP files bug when reinstalling plugin
When reinstalling or uninstalling plugin, some of the EnergyPlus files can escape deletion, meaning that the directory persists. This would previously mean the EP installation was skipped but there would be missing files. Now, a check is done to verify that all required files are present - if not, the entire EP directory is deleted and sourced again.